Put JPCalibHandler to utils.correction
